Move specification tests to folder
